+++ packages/backend-function/src/schedule_parser.test.ts
@@ -1,194 +1,302 @@
 import { describe, it } from 'node:test';
 import { App, Duration, Stack } from 'aws-cdk-lib';
 import { Code, Function, Runtime } from 'aws-cdk-lib/aws-lambda';
-import { convertFunctionSchedulesToScheduleExpressions } from './schedule_parser.js';
+import { convertFunctionSchedulesToScheduleProps } from './schedule_parser.js';
 import { CronSchedule, FunctionSchedule } from './factory.js';
 import assert from 'node:assert';
 import os from 'os';
 
 void describe('ScheduleParser', () => {
-  void it('creates EventBridge Schedule expression for natural language', () => {
+  void it('creates EventBridge Schedule props for natural language', () => {
     const schedule: FunctionSchedule[] = ['every day'];
     const expectedScheduleExpression = 'cron(0 0 * * ? *)'; // every day at 00:00
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       schedule,
     );
 
-    assert.equal(schedules.length, 1);
-    assert.equal(schedules[0].expressionString, expectedScheduleExpression);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps.length, 1);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedScheduleExpression,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expression for natural language with timezone', () => {
+  void it('creates EventBridge Schedule props for natural language with timezone', () => {
     const schedule: FunctionSchedule[] = [
       { rate: 'every day', timezone: 'America/New_York' },
     ];
     const expectedScheduleExpression = 'cron(0 0 * * ? *)'; // every day at 00:00
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       schedule,
     );
 
-    assert.equal(schedules.length, 1);
-    assert.equal(schedules[0].expressionString, expectedScheduleExpression);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'America/New_York');
+    assert.equal(scheduleProps.length, 1);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedScheduleExpression,
+    );
+    assert.equal(
+      scheduleProps[0].schedule.timeZone?.timezoneName,
+      'America/New_York',
+    );
+    assert.equal(scheduleProps[0].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expression for natural language with number value', () => {
+  void it('creates EventBridge Schedule props for natural language with description', () => {
+    const schedule: FunctionSchedule[] = [
+      { rate: 'every day', timezone: 'UTC', description: 'Daily task' },
+    ];
+    const expectedScheduleExpression = 'cron(0 0 * * ? *)'; // every day at 00:00
+    const testLambda = getTestLambda();
+
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
+      testLambda,
+      schedule,
+    );
+
+    assert.equal(scheduleProps.length, 1);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedScheduleExpression,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, 'Daily task');
+  });
+
+  void it('creates EventBridge Schedule props for natural language with number value', () => {
     const schedule: FunctionSchedule[] = ['every 5m'];
     const expectedScheduleExpression = 'cron(*/5 * * * ? *)';
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       schedule,
     );
 
-    assert.equal(schedules.length, 1);
-    assert.equal(schedules[0].expressionString, expectedScheduleExpression);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps.length, 1);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedScheduleExpression,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expression for cron expression', () => {
+  void it('creates EventBridge Schedule props for cron expression', () => {
     const schedule: FunctionSchedule[] = ['* * * * ?'];
     const expectedScheduleExpression = 'cron(* * * * ? *)';
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       schedule,
     );
 
-    assert.equal(schedules.length, 1);
-    assert.equal(schedules[0].expressionString, expectedScheduleExpression);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps.length, 1);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedScheduleExpression,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expression for cron expression with timezone', () => {
+  void it('creates EventBridge Schedule props for cron expression with timezone', () => {
     const schedule: FunctionSchedule[] = [
       { cron: '* * * * ?', timezone: 'America/New_York' },
     ];
     const expectedScheduleExpression = 'cron(* * * * ? *)';
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       schedule,
     );
 
-    assert.equal(schedules.length, 1);
-    assert.equal(schedules[0].expressionString, expectedScheduleExpression);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'America/New_York');
+    assert.equal(scheduleProps.length, 1);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedScheduleExpression,
+    );
+    assert.equal(
+      scheduleProps[0].schedule.timeZone?.timezoneName,
+      'America/New_York',
+    );
+    assert.equal(scheduleProps[0].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expression for cron expression with /', () => {
+  void it('creates EventBridge Schedule props for cron expression with description', () => {
+    const schedule: FunctionSchedule[] = [
+      {
+        cron: '* * * * ?',
+        timezone: 'America/New_York',
+        description: 'Every minute',
+      },
+    ];
+    const expectedScheduleExpression = 'cron(* * * * ? *)';
+    const testLambda = getTestLambda();
+
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
+      testLambda,
+      schedule,
+    );
+
+    assert.equal(scheduleProps.length, 1);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedScheduleExpression,
+    );
+    assert.equal(
+      scheduleProps[0].schedule.timeZone?.timezoneName,
+      'America/New_York',
+    );
+    assert.equal(scheduleProps[0].description, 'Every minute');
+  });
+
+  void it('creates EventBridge Schedule props for cron expression with /', () => {
     const schedule: FunctionSchedule[] = ['*/5 * * * ?'];
     const expectedScheduleExpression = 'cron(*/5 * * * ? *)';
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       schedule,
     );
 
-    assert.equal(schedules.length, 1);
-    assert.equal(schedules[0].expressionString, expectedScheduleExpression);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps.length, 1);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedScheduleExpression,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expression for cron expression with ,', () => {
+  void it('creates EventBridge Schedule props for cron expression with ,', () => {
     const schedule: FunctionSchedule[] = ['0 1,2 * * ?'];
     const expectedScheduleExpression = 'cron(0 1,2 * * ? *)';
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       schedule,
     );
 
-    assert.equal(schedules.length, 1);
-    assert.equal(schedules[0].expressionString, expectedScheduleExpression);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps.length, 1);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedScheduleExpression,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expression for cron expression with -', () => {
+  void it('creates EventBridge Schedule props for cron expression with -', () => {
     const schedule: FunctionSchedule[] = ['0 1-5 * * ?'];
     const expectedScheduleExpression = 'cron(0 1-5 * * ? *)';
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       schedule,
     );
 
-    assert.equal(schedules.length, 1);
-    assert.equal(schedules[0].expressionString, expectedScheduleExpression);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps.length, 1);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedScheduleExpression,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expressions for an array of natural language', () => {
+  void it('creates EventBridge Schedule props for an array of natural language', () => {
     const functionSchedules: FunctionSchedule[] = ['every 2h', 'every month'];
     const expectedExpressionTwoHours = 'cron(0 */2 * * ? *)';
     const expectedExpressionEveryMonth = 'cron(0 0 1 * ? *)';
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       functionSchedules,
     );
 
-    assert.equal(schedules.length, 2);
-    assert.equal(schedules[0].expressionString, expectedExpressionTwoHours);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'UTC');
-    assert.equal(schedules[1].expressionString, expectedExpressionEveryMonth);
-    assert.equal(schedules[1].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps.length, 2);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedExpressionTwoHours,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, undefined);
+    assert.equal(
+      scheduleProps[1].schedule.expressionString,
+      expectedExpressionEveryMonth,
+    );
+    assert.equal(scheduleProps[1].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[1].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expressions for an array of cron expressions', () => {
+  void it('creates EventBridge Schedule props for an array of cron expressions', () => {
     const functionSchedules: FunctionSchedule[] = ['* * * * ?', '0 0 * * ?'];
     const expectedExpressionEveryMinute = 'cron(* * * * ? *)';
     const expectedExpressionEveryMidnight = 'cron(0 0 * * ? *)';
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       functionSchedules,
     );
 
-    assert.equal(schedules.length, 2);
-    assert.equal(schedules[0].expressionString, expectedExpressionEveryMinute);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps.length, 2);
     assert.equal(
-      schedules[1].expressionString,
+      scheduleProps[0].schedule.expressionString,
+      expectedExpressionEveryMinute,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, undefined);
+    assert.equal(
+      scheduleProps[1].schedule.expressionString,
       expectedExpressionEveryMidnight,
     );
-    assert.equal(schedules[1].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[1].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[1].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expressions for an array of both natural language and cron expressions', () => {
+  void it('creates EventBridge Schedule props for an array of both natural language and cron expressions', () => {
     const functionSchedules: FunctionSchedule[] = ['* * * * ?', 'every week'];
     const expectedExpressionEveryMinute = 'cron(* * * * ? *)';
     const expectedExpressionEveryWeek = 'cron(0 0 ? * 1 *)';
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       functionSchedules,
     );
 
-    assert.equal(schedules.length, 2);
-    assert.equal(schedules[0].expressionString, expectedExpressionEveryMinute);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'UTC');
-    assert.equal(schedules[1].expressionString, expectedExpressionEveryWeek);
-    assert.equal(schedules[1].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps.length, 2);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedExpressionEveryMinute,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, undefined);
+    assert.equal(
+      scheduleProps[1].schedule.expressionString,
+      expectedExpressionEveryWeek,
+    );
+    assert.equal(scheduleProps[1].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[1].description, undefined);
   });
 
-  void it('creates EventBridge Schedule expressions for an array of both expression with timezone and cron expression without timezone', () => {
+  void it('creates EventBridge Schedule props for an array of both expression with timezone and cron expression without timezone', () => {
     const functionSchedules: FunctionSchedule[] = [
       '* * * * ?',
       { cron: '0 0 * * ?', timezone: 'Asia/Tokyo' },
@@ -197,19 +305,56 @@ void describe('ScheduleParser', () => {
     const expectedExpressionEveryMidnight = 'cron(0 0 * * ? *)';
     const testLambda = getTestLambda();
 
-    const schedules = convertFunctionSchedulesToScheduleExpressions(
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
       testLambda,
       functionSchedules,
     );
 
-    assert.equal(schedules.length, 2);
-    assert.equal(schedules[0].expressionString, expectedExpressionEveryMinute);
-    assert.equal(schedules[0].timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps.length, 2);
     assert.equal(
-      schedules[1].expressionString,
+      scheduleProps[0].schedule.expressionString,
+      expectedExpressionEveryMinute,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, undefined);
+    assert.equal(
+      scheduleProps[1].schedule.expressionString,
       expectedExpressionEveryMidnight,
     );
-    assert.equal(schedules[1].timeZone?.timezoneName, 'Asia/Tokyo');
+    assert.equal(
+      scheduleProps[1].schedule.timeZone?.timezoneName,
+      'Asia/Tokyo',
+    );
+    assert.equal(scheduleProps[1].description, undefined);
+  });
+
+  void it('creates EventBridge Schedule props for an array with descriptions', () => {
+    const functionSchedules: FunctionSchedule[] = [
+      { rate: 'every 2h', timezone: 'UTC', description: 'Hourly sync' },
+      { cron: '0 0 * * ?', timezone: 'UTC', description: 'Daily cleanup' },
+    ];
+    const expectedExpressionTwoHours = 'cron(0 */2 * * ? *)';
+    const expectedExpressionEveryMidnight = 'cron(0 0 * * ? *)';
+    const testLambda = getTestLambda();
+
+    const scheduleProps = convertFunctionSchedulesToScheduleProps(
+      testLambda,
+      functionSchedules,
+    );
+
+    assert.equal(scheduleProps.length, 2);
+    assert.equal(
+      scheduleProps[0].schedule.expressionString,
+      expectedExpressionTwoHours,
+    );
+    assert.equal(scheduleProps[0].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[0].description, 'Hourly sync');
+    assert.equal(
+      scheduleProps[1].schedule.expressionString,
+      expectedExpressionEveryMidnight,
+    );
+    assert.equal(scheduleProps[1].schedule.timeZone?.timezoneName, 'UTC');
+    assert.equal(scheduleProps[1].description, 'Daily cleanup');
   });
 
   void it('throws if rate is a negative number', () => {
@@ -218,7 +363,7 @@ void describe('ScheduleParser', () => {
 
     assert.throws(
       () => {
-        convertFunctionSchedulesToScheduleExpressions(testLambda, schedule);
+        convertFunctionSchedulesToScheduleProps(testLambda, schedule);
       },
       {
         message:
@@ -233,7 +378,7 @@ void describe('ScheduleParser', () => {
 
     assert.throws(
       () => {
-        convertFunctionSchedulesToScheduleExpressions(testLambda, schedule);
+        convertFunctionSchedulesToScheduleProps(testLambda, schedule);
       },
       {
         message:
@@ -257,7 +402,7 @@ void describe('ScheduleParser', () => {
 
     assert.throws(
       () => {
-        convertFunctionSchedulesToScheduleExpressions(testLambda, schedule);
+        convertFunctionSchedulesToScheduleProps(testLambda, schedule);
       },
       {
         message: expectedErrors.join(os.EOL),
@@ -284,7 +429,7 @@ void describe('ScheduleParser', () => {
 
     assert.throws(
       () => {
-        convertFunctionSchedulesToScheduleExpressions(testLambda, schedule);
+        convertFunctionSchedulesToScheduleProps(testLambda, schedule);
       },
       {
         message: expectedErrors.join(os.EOL),
@@ -307,7 +452,7 @@ void describe('ScheduleParser', () => {
 
     assert.throws(
       () => {
-        convertFunctionSchedulesToScheduleExpressions(testLambda, schedule);
+        convertFunctionSchedulesToScheduleProps(testLambda, schedule);
       },
       {
         message: expectedErrors.join(os.EOL),
@@ -330,7 +475,7 @@ void describe('ScheduleParser', () => {
 
     assert.throws(
       () => {
-        convertFunctionSchedulesToScheduleExpressions(testLambda, schedule);
+        convertFunctionSchedulesToScheduleProps(testLambda, schedule);
       },
       {
         message: expectedErrors.join(os.EOL),
@@ -344,7 +489,7 @@ void describe('ScheduleParser', () => {
 
     assert.throws(
       () => {
-        convertFunctionSchedulesToScheduleExpressions(testLambda, schedule);
+        convertFunctionSchedulesToScheduleProps(testLambda, schedule);
       },
       {
         message:
@@ -362,7 +507,7 @@ void describe('ScheduleParser', () => {
 
     assert.throws(
       () => {
-        convertFunctionSchedulesToScheduleExpressions(testLambda, schedule);
+        convertFunctionSchedulesToScheduleProps(testLambda, schedule);
       },
       {
         message: "Could not determine the function's schedule type",