+++ packages/plugin-types/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aws-amplify/plugin-types
 
+## 1.11.2
+
+### Patch Changes
+
+- 7d0ba5e: chore: upgrade CDK dependencies
+- 4603f7a: bump aws-cdk-lib version to ^2.234.1 across all packages
+
 ## 1.11.1
 
 ### Patch Changes